requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
